com/pyproject.toml
[project]
name = "com-service"
version = "0.1.0"
description = "Context Orchestrator Management for QA Intelligence"
authors = [{name = "QA Intelligence Team"}]
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "faiss-cpu>=1.7.4",
    "sentence-transformers>=2.2.2",
    "numpy>=1.24.3",
    "pyarrow>=14.0.1",
    "openai>=1.3.5",
    "anthropic>=0.7.7",
    "whoosh>=2.7.4",
    "networkx>=3.2.1",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.1",
    "click>=8.1.7",
    "rich>=13.7.0",
    "httpx>=0.25.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "black>=23.12.1",
    "ruff>=0.1.8",
]

[project.scripts]
com = "cli.main:cli"
com-server = "api.main:start"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
